spawning a little different now particles start with a faster velocity inside a galaxy galaxies depth's are now 1/100 of radius
git-svn-id: svn://gakman.com/gravit/trunk@150 60970b36-f649-e149-802f-859a9d0a703e

+++ gfx.c
@@ -353,7 +353,7 @@ void drawFrame() {
 			if (screen[2] > 1)
 				continue;
 
-			size = view.particleSizeMin + (1.f - screen[2]) * view.particleSizeMax;
+			size = view.particleSizeMin + (1.f - (float)screen[2]) * view.particleSizeMax;
 
 			glBegin(GL_QUADS);
 			glColor4fv(pd->col);
